# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a Web Push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionBody"
      responses:
        "200":
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/unsubscribe:
    delete:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a Web Push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeBody"
      responses:
        "200":
          description: Subscription removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/subscription-status:
    get:
      operationId: getPushSubscriptionStatus
      tags: [push]
      summary: Check whether the current user has any active push subscription
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionStatus"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/prefs:
    get:
      operationId: getPushPrefs
      tags: [push]
      summary: Get current user's push notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushPrefs"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePushPrefs
      tags: [push]
      summary: Update push notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushPrefsUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushPrefs"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PushSubscriptionKeys:
      type: object
      properties:
        p256dh:
          type: string
        auth:
          type: string
      required:
        - p256dh
        - auth
    PushSubscriptionBody:
      type: object
      properties:
        endpoint:
          type: string
          format: uri
        keys:
          $ref: "#/components/schemas/PushSubscriptionKeys"
      required:
        - endpoint
        - keys
    PushUnsubscribeBody:
      type: object
      properties:
        endpoint:
          type: string
          format: uri
      required:
        - endpoint
    PushSubscriptionStatus:
      type: object
      properties:
        subscribed:
          type: boolean
      required:
        - subscribed
    PushPrefs:
      type: object
      properties:
        pushExpiry:
          type: boolean
        pushBroadcast:
          type: boolean
      required:
        - pushExpiry
        - pushBroadcast
    PushPrefsUpdate:
      type: object
      properties:
        pushExpiry:
          type: boolean
        pushBroadcast:
          type: boolean
...